SHOW STATE command
This new command exposes PgBouncer states which are active, pause or
suspend (see PauseMode). Although it is possible to obtain this state
looking at SHOW DATABASES, you have to process multiple entries to
discover the global state; it seems a fragile solution. PgBouncer has an
internal variable that knows the current state so let's expose it.

It shows one state per row so we have some room for additional states in
the future.

Closes #528