Add timestamp to SHOW commands
Postgres has a facility to include the function now() while collecting
statistics from it. Unfortunately PgBouncer does not have this option.
When we are collecting statistics from PgBouncer, the collector
generally obtain the current timestamp and attach this information to
the statistics that it was collected. Instead of relying on this
additional step, this change includes a new column (timestamp) that
already provides the timestamp.